# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedRushApp API",
        "description": "API REST para el sistema de distribución de medicinas MedRush",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Servidor Local"
        },
        {
            "url": "https://api.medrush.com",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
...
        "/api/users/{userId}/activo": {
            "patch": {
                "tags": [
                    "User",
                    "Base"
                ],
                "summary": "Activar o desactivar usuario",
                "description": "Activa o desactiva un usuario. Los administradores no pueden desactivarse a sí mismos ni puede haber solo uno activo en el sistema.",
                "operationId": "userToggleActive",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "UUID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nuevo estado activo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success"
                                            ],
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El estado activo de la cuenta ha sido actualizado."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se puede desactivar el último administrador activo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Valor inválido para is_active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{userId}/avatar": {
            "post": {
                "tags": [
                    "User",
                    "Base"
                ],
                "summary": "Subir o actualizar foto de perfil",
                "description": "Carga o actualiza la foto de perfil del usuario. Puede omitirse el archivo para eliminar la foto actual.",
                "operationId": "userUploadPicture",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "UUID del usuario",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Archivo de imagen",
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "description": "Imagen JPEG/PNG. Opcional - si se omite, elimina la foto actual",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto de perfil actualizada o eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success"
                                            ],
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foto de perfil actualizada correctamente."
                                        },
                                        "data": {
                                            "properties": {
                                                "url": {
                                                    "description": "URL firmada para descargar la imagen",
                                                    "type": "string",
                                                    "format": "url",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Archivo inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al procesar la imagen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "description": "Estructura base de error en la respuesta",
                "properties": {
                    "code": {
                        "description": "Código de error único",
                        "type": "string",
                        "example": "VALIDATION_ERROR"
                    },
                    "errors": {
                        "description": "Mapa de errores por campo (opcional)",
                        "type": "object",
                        "example": {
                            "email": [
                                "Las credenciales proporcionadas son incorrectas."
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "BadRequestResponse": {
                "description": "Respuesta 400 - Solicitud malformada",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "fail"
                        ],
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "La solicitud no pudo ser procesada debido a una sintaxis inválida o datos incorrectos."
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "description": "Respuesta 401 - No autenticado",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "fail"
                        ],
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "No autenticado. Por favor, inicie sesión."
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "string",
                                "enum": [
                                    "UNAUTHORIZED"
                                ],
                                "example": "UNAUTHORIZED"
                            },
                            "errors": {
                                "type": "null"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "description": "Respuesta 422 - Error de validación en los datos",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "fail"
                        ],
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "Los datos proporcionados no son válidos."
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "string",
                                "enum": [
                                    "VALIDATION_ERROR"
                                ],
                                "example": "VALIDATION_ERROR"
                            },
                            "errors": {
                                "description": "Errores de validación por campo",
                                "type": "object",
                                "example": {
                                    "email": [
                                        "Email es requerido"
                                    ],
                                    "password": [
                                        "Password debe tener al menos 8 caracteres"
                                    ]
                                },
                                "additionalProperties": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ServerErrorResponse": {
                "description": "Respuesta 500 - Error interno del servidor",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "error"
                        ],
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error interno del servidor."
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "string",
                                "enum": [
                                    "SERVER_ERROR"
                                ],
                                "example": "SERVER_ERROR"
                            },
                            "errors": {
                                "type": "null"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de acceso Sanctum. Se obtiene en el endpoint de login.",
                "name": "Token based based auth",
                "in": "header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}